Refactored tree generation code to separate files
